dashboard/requirements.txt
matplotlib==3.8.1
numpy==1.26.4
pandas==2.2.1
python_aqi==0.6.1
seaborn==0.13.2
streamlit==1.31.1
